ArchiveMa: list compressed archive contents, Viewer gzip
listContents of ArchiveMa has been updated to first check if the
file passed to it is compressed or not, and if compressed get a
buffered input into the decompressed data. And inturn either the
original buffered input or this new buffered input is what is
finally passed to the archiver logic to extract the info about
the contents within it.

ViewerActivity has been updated to handle gzip mime type.