Handle BZIP2 and non-mimetyped archive files - P1
For some reason android/java's net uri mimetype helper is not identifying
mimetype for bzip2 files. Also similar issue could be there for other
archive files and or one may forget to specify the mimetype.

So now additional file extension based mapping into archive file type
logic is used to check if the file at hand is compressed and or archived,
if so it will be passed to handleArchive.

This will also take care of AR, ARJ etc types, which I am not checking for
as part of mimetype check, but am identifying in the ext2archive helper.