Handle 7z and xz files
As support for xz compression be it a tar.xz file or a 7z file with
xz and related compressions (lzma/...), depends on the org.tukaani.xz
library, from apache commons-compress library perspective, so the
same has been added as a build dependency to the project.

showArchive has been updated to take a argument which is the archive/
compression type. In turn if it finds that it is a 7z file, then it
calls the 7z related listArchive7z function, else generic listArchive
function is called.

Added logic to convert from the mimetype to corresponding ArchiveMa's
type identifiers for the archive/compressed files. Which inturn is
passed to the updated showArchive function.

TODO: Rather I should be able to handle the difference between 7z
and other archive formats in the generic listArchive function, which
I have. Rather look at the exception raised when trying to use 7z
with ArchiveStreamFactory, and then switch to the 7z related logic.