Viewer: Kotlin TOCHECK wrt sFiles, return true 4 longClick
For some reason kotlin didnt allow the syntax $sFiles[it] in onClick,
but allowed it in onLongClick. And inturn at runtime $sFiles[it] dumped
the full $sFiles ArrayList, while ${sFiles?.get(it)} prints only the
specified item in the ArrayList. Need to look at Kotlin bit more wrt
this.

As not returning true for long click will trigger a normal click,
immidiately after longClick handler returns, so now returning true
in the long click handler to avoid this.