ArchiveMa:P1: listEntries handle 7z indirectly
Trying to make listEntries handle 7z indirectly by instantiating 7z
or calling into 7z logic.

NOTE: Here there is a var on a outer block of code and val in the inner
block of the code both having the same name, so once the inner block
of code is exited the value accesible will be the one corresponding to
the outer block. And here the idea was not that. Rather the idea was
to move that val to the outer block and make it a var, but forgot to
remove the val in front of the inner assignment. BE CAREFUL TO AVOID
SUCH OVERSIGHTS.