ArchiveMa:P2: Handle 7z indirectly from listEntries
As 7z class doesnt return a archive input stream, so on exception
identifying 7z file, I cant directly instantiate the 7z and inturn
use a common block to iterate over the entries (That was the initial
idea). So instead it calls into the listEntries7z on exception related
to 7z in the listEntries archive handling block.

Not clean but handles the situation, if user forgets to call listEnt7z
and instead calls the normal listEnt(ries) wrongly.