ArchiveMa:Kotlin: Cant revert to this partial self
As the try is a new inner block, so a var or val declared in it cant
be used outside that inner block. Unlike C, where a local variable
declared inbetween the code remains in context till that function ends
and is not restricted to just the inner block in the func where its
declared. IE in C the location of the declaration of the local variable
is just from a end user reading perspective, while from a compiler
perspective it is as good as declaring at the begining of the func.

NOTE: Will be reverting this change next, just captured here to show
the issue, to myself, so that at a later date, if I revisit, I can
remember kotlin semantic from the log, instead of going thro the lang
docs.