Revert "ArchiveMa:Kotlin: Cant revert to this partial self"
This reverts commit bbc79053bc90359e5a67504593a3fbf8ce4b5868.

Reverting as mentined in the last commit. Also I think a var with
null can have close called potentially and or more potentially one
can call close multiple times on a stream handle.

Also even thou a variable/val assigned in a inner block cant be
accessed from outside that inner code block by the developer. As
far as the runtime is concerned, the inner var or val remains in
existance and actively participating till the last reference to
it is removed (I assume, assuming kotlin uses garbage collection).
And or the outer major code block like function exits(i.e returns
to its caller).

Otherwise the inFileC which feeds data into inFileX (in case the
file is a archive which is compressed) would have been a issue.