SimpRecycView: Click handlers pass view also now
SimpRecycView updated to pass the view which received the click event
also to the listeners of both normal and long click. Previously they
were only passing the position of the data item corresponding to the
view, in the overall datalist associated with the SimpRecycView.

In turn Viewer activity updated to account for the same in its
call back handlers for these events.