SimpRecycView: handle multiselection list by default
Added logic to maintain a multiple item selection list on its own
by default. THe user can always disable this logic by setting
bHandleMultiSelection to false.

If enabled, then when ever user does a long click or press, the
corresponding item is toggled wrt the selected list. i.e if it is
not yet selected, then it is selected. However if it is already
selected, then it is deselected.

Inturn when binding a selected item into the view holder, the
corresponding itemview's isactivated state is set. Else it will
be cleared.

When user assigns a new dataList, then selected list is cleared.