SimpRecycView: Breathing space around itemview, Java 1.6
As the internal default textview was being used in the layout with
out any margin around it, it was appearing too dense, so have added
some margin around/wrt each itemview.

In the process, found two functions to set the margin for textview,
one which takes margin for each side of the view seperately and
another which takes a single value and internally sets up the 4
individual values. As initially I wanted to experiment with same
value for all 4 sides, so tried the func with single arg, but then
it raised a compile time error telling trying to mix bytecode for
jvm target 1.8 with code for jvm target 1.6.

    So in project structure related setting made both the source
    and target compatibility to be explicitly 1.6 (Java 6).
    Also used the age old setMargin with 4 args.

    The idea being I want to be compatible with jvm target 1.6
    for now.