SRcV: SRCVBindView function type and variable
Added a variable to store the optional callback to be called when
the GUI and inturn RecyclerView wants to show a item on the screen
by mapping it to a existing view in its cache of views.

NOTE: Initially I had planed on exposing the viewholder in the callback
similar to how RecyclerView behaves, but have now instead decided
to expose only the view created by the developer, as view holder is
mainly internal to SRcV.

TODO: Even SRCVCreateViewHolder is expected to return only a view
and SRcV internally maps it to a internal View Holder. Will be
renaming it to match its functionality better.