SRcV: Constructor with columnCount, Updated initHelper
Instead of hard coding the use of defaultColumnCount, rather now
the user could explicilty call a new constructor of SRcV which
allows them to specify the columnCount.

If a constructor without columnCount is called or if columnCount is
set to -1, then it fallsback to the defaultColumnCount.

When inflating from xml layout resource, the constructor called
doesnt accept columnCount, so in this flow it fallsback to defaultCC.

initHelper updated to provide the required behaviour wrt columnCnt
and its fallback into defaultColumnCount(CC).