SRcV: onKeyListener for spacebar for selection
TO allow end users who use a keyboard to work with SRcV, to be able
to select items as part of multi selection, now onKeyListener is
trapped wrt the item views.

Moved the logic in onLongClick into a helper func, which is now inturn
called from both onLongClikc as well as onKey.

TODO: Need to rename previous handleSelection to maybe _handleSelection,
as it basically handles only a part of the required logic. Or maybe
even handleSelectedPlus.