SRcV: focus, assignDataList initialPos; FManFrag simplified
As users of SRcV may want a given item in the list/grid to be in
focus for different reasons like

a) A default item other than the 1st item to be in focus, when a new
data list is loaded into it. NOTE: Rather by default no item will be
in focus and as user interacts with the GUI, the nearest item will
get focus and on key board navigation by default it will start at the
top item, if coming from top. But many a times we may want to change
this behaviour.

b) based on some other end user interaction and or other programmatic
reasons set a specific item to be having focus.

So added a focus helper routine, which tries to set the specified
position in the item list to have focus.

Also assignDataList updated to take a initialPosition argument,
which tells which item in the list should have focus, after the
new data has been updated into the SRcV list/grid.