SRcV: Moved scrollToPosition to focus
As sometimes the position in the SRcV requested to get focus may
not be visible in the first place, so the logic first tries to
get the position to be in view by scrolling if required. ANd only
then it tries to request focus, through a posted runnable, so that
scrolling can finish its job first, before focus is requested,
else wrong item may get focus and or it may not get focus. Bcas
scrolling can change the mapping of item views to the corresponding
item data (bcas a small cache of views is used/reused to provide an
efficient list/grid view)