SRcV: parentViewGroup for SRCVCreateView, FManFrag uses
During inflating to allow layout parameters to flow through into
the item view, and or to allow item view to be a child of the
higher up viewgroup in the ui heierarchy, now parent viewgroup
got from onCreateViewHolder is passed on to the SRCVCreateView.

Its upto a given user to decide if they want to use it or not.

FManFrag updated to use SRCVCreateView to setup its custom item
view.