Main:ContextMenu: initial skeletal menu preperation
Added code to maintain a state for the main activity.
It consists of Normal, CanPaste and Transfering for now.

Based on the state and availability of selected items if any,
a context menulist is created.

To help with the logic, for now SRcV of FManFrag has been made
public. Also bcas without Fragment logic, eitherway it would have
been part of the MainActivity, so end result no change at one lvl.