SRcV: textView padding instead of margins, dp->px
To make the view uncluttered previously I was adding margin around
each textview if item view is internally managed. But as margin
is considered to be outside the textview, so currently I am not sure
if user interaction in this area will be considered as part of the
corresponding item view's or ignored. Also the highlighting is also
restricted to the text part and margin is not included.

So for these reasons, now I use padding instead of margin, because
padding is supposed to be included as part of the view it belongs
to unlike the margin.

Also to scale appropriately between different displays, now I treat
the hardcoded values as dp (i.e density independent pixel values)
and inturn convert them to pixels using android's density scaling
factor, when assigning to margin/padding.