ContextMenu: cleanup of initial run/experiment
As user may want to go up/back in the directory heirarchy, provide
it has the 1st option, as they might have hit this button instead
of pressing back.

Cleanup the logic which decides whether to show the copy menu option
or not, yesterday night I had just put the basic skeleton and not
done it fully (i.e show only if there are some selected contents).