CopyPaste: Handle Copy, log Paste, Avoid CanPaste state
As the list of selected files till now needs to be stored somewhere,
the same is now used to decide whether there is anything to paste.
Thus avoiding the need for CANPASTE state, which is redundant and
prime for out-of-sync if it is not kept in sync with the list.

Copy logic has been implemented, it copies the currently selected
list of files into the global selectedFilesList. This allows one
to have files from different directories copied into a given
destination at the same time.

TODO:TOTHINKLATER:Or should I reset the global selectedFilesList
if a new set of files is added from a different directory?

Add exit to the context menu.