ContextMenu: lookat android PopupMenu not alertDialog
Looking at using android's menu and inturn popup menu support to
handle the application context menu.

Ideally would have used the menu linked to the application appbar.
However on android tv, there is no default support for appbar. So
also why I originally used the alert dialog, but as it occupies
too much screen space, so moving to the popup menu.

However using sub menu seems to be leading to a exception in the
android tv api level 28 avd. There seems to be similar issue noticed
by others also from reading on internet. It seems to be related to
having a header/title in the menu.

So will be reverting to single menu rather than a heirarchy, even
thou one of the reasons why I originally looked at menus' was bcas
of the possibility of menu heirarchy.