MaContextMenu: Handle few items, Remove submenu, reset copy
Have removed the transfer submenu and put its items at the top lvl
itself, for reasons noted in last commit.

Have removed the contextMenuDlg and have moved/implemented some of
the logic related to menu selection now in the popup logic.

Have provided the developer option to decide, whether user can
add-to copy / send list from across different directories or is
it restricted to the last list of selected items belonging to a
given directory.