MaContextMenu: Disable rather than Hide MenuItems
For actions which cant be done at the given time, like if there are
no selected items in the current directory/path, then we cant do
copy or send; for such actions we set them to disabled rather than
hidden.

This aids discovery of features, by showing to user that such a
functionality is available, but that currently it cant be triggered.