CoRoutines: Experimenting to get a feel
Created a coroutine to try and run long running io task on a seperate
io thread instead of the main ui thread.

At the same time in the main ui thread, if btnMa is clicked, it checks
if any previous io job is still active and then asks user to wait.