Paste: better copyRecursive, err info in btnMa
Now copyRecursive sets the destination to be a file, if the given
dest is a directory and the source is a file. This should potentially
help wrt this corner case wrt java File().copyRecursively.

Also copyRecursive returns success or failure related boolean, as well
as a list of files which failed copy.

Previously it would have always returned success, as onError handler
always returns SKIP and not TERMINATE. Now if onError is triggered,
then copyRecursive will return false.

Now btnMa text is updated with errCnt if it is not zero.