Flexible handleFileIO and Copy,Delete
Added Delete menu option.

Modified handlePaste to handleFileIO, which inturn takes the kind
of fileio to do. It currently supports copy and delete.

Inturn paste and delete menu items now inturn trigger this new
function, with the argument being either copy or delete.

The reason handleFileIO was created is because most file io are
io intensive operations which may take some time, so should be
run on a different thread and not the main ui thread. ANd thus
this logic can be shared across file io operations.