handleFileIO and handleDelete
handleFileIO updated to take the list of files it should operate
on. In case of pasting it will be the global selectedFilesList.
In case of deleting it will be the local selectedList in menu
handler. Bcas of this difference this change was needed to handle
FileIO function. Also it makes the handleFileIO func independent
at one level.

NOTE: However it still has the sideeffect of clearing the global
selectedFilesList, if it is a copy operation.

Added a handleDelete logic, so that it can create the list of
file path strings from the selectedList of FManItems.