handleFileIO refreshes MainActivity loadPath view
As fileio operations of copy or delete will change the contents
of the current directory (normal case) in which the view was, when
the operation was started, so it will trigger a loadPath (with no
arg so that curPath will get loaded) to update the view.

If the user had not changed the path when the fileio was occuring,
then it will be refreshed and the new update content will be shown.

If the user had changed into some other path, then the current
view corresponding to that will be refreshed and shown again(which
is harmless, except for selections in current view, which will get
cleared).