HPopupMenu: curPath getter, call stack p1
Now curPath is derived when someone uses it by using its constituent
parts, which are both part of the class instance. The constituent
parts i.e curLvl and curId are kept updated as required, when one
is navigating.

Added a call stack, into which curPath is pushed when navigating in.
Later it will be poped as one is navigating back.