FMan:Copy: Append source file/dir name to dst dir
If the destination specified to copyRecursive is a directory, then
append the source file or directory name to it, so that the source
file or directory gets copied into the destination and still retain
its identity.

Previoulsy if file was being copied into a directory, it will be fine.
But if a directory is being copied into a directory, rather than the
directory being copied, instead all its contents is what would get
copied.

Now the source identity (be it a file or directory) is retained
with-in the destination directory specified.

If one wants the old behaviour, they will have to go in to the
required source directory and select the files within and then
do a copy followed by the paste.