Main: implement handleNewFolder logic
A alert dialog which takes a single line text, with a cancel and
create button is used to get the folder name to create and then
the same is created in the current folder, if requested.

By default the inputType enables multiline text. So reseting it to
a simple text type (i.e without oring with multiline), avoids the
multiline support of edittext.