Main:MaContextFileIO failure BtnMa alert constant
If there is a file io related failure as part of any of the MaContext
Menu file io operations, then there is a failure alert shown to the
end user by changing the btnMa text.

THis inturn gets reset back to the original BtnMa text after few
seconds. The number of seconds involved has been declared as a
constant value now, instead of hardcoding in different places.