Main: Call ExternalViewer, even if Mimetype unknown for file
Renamed the variable which decides whether Internal viewer should
be called before external viewer to bViewFileInternalFirst.

With last commit, now end user can control this variable. Previously
it was configurable at compile time only.

Now even if mimetype of the file is not known, the external viewer
will be called, if required. Previously external viewer was called
only if the mimetype is known. THis was to avoid generic trap all
activities from trying to get hold of the file.

NOTE THAT if the internal viewer returns with a failure, we chain
into external viewer. However if external viewer returns with a
failure, we dont chain into internal viewer. This is because, most
external viewers dont consistently return with success or failure
to view file related status.