Main: SaveRestore bViewFileInternalFirst
Now if activity is stopped and resumed due to some other app gaining
foreground focus and or because of system configuration change like
rotation of screen etc, the user selection of bViewFileInternalFirst
is saved and restored as required.

TODO: Need to save this across full application stop and fresh starts
by using shared preference or some other such mechanism.