SRcV: Touch->LongClick - NO -> Click
If a touch event has already triggered a LongClick, because the end
user has continued touching for a long time, then dont trigger a
Click when the touch is released after the long touch.

Otherwise it was leading to wrong behaviours.

NOTE: However it is noticed that the recycler view is not clearing
the activated state properly, even thou notifyDataSetChanged is
called. AND OR there is some corner case in my flow.