SRcV: Touch and View state update Visualisation fix
Trying to just do a invalidate or postInvalidate or even doing
postInvalidateDelayed_1sec, didnt help with itemviews strangly
having wrong state (i.e activated even if they are not activated)
get fixed. But at same time triggering a popupMenu would clear
this wrong states at the same time.

This issue was not there originally when testing on Emulator as
well as later on a actual mobile. But then in mobile is when I
realised that touch was not triggering things in the normal touch
flow perspective. Instead touch was working like a mouse click,
by making the 1st touch act like a change focus and then a 2nd
touch triggering the click event. So to fix this mouse click kind
of flow and get a proper touch based flow working, I added logic
to trap onTouch and handle it suitably. ANd that is when this
wrong state based itemview background drawable visual bug came in.

TO fix this, next I tried calling remove all views through layout
manager and that seems to have helped. Parallely I have retained
the postInvalidateDelayed and reduced it to 0.5 sec delay. Havent
yet tested having only remove all views and not postInvalidateDelayed.