Background+CoRoutine based initial path loading and display
Previously FManFrag as part of its onCreateView was trying to load
the default path file list from storage volume/disk and inturn
update the same into its SRcV.

In turn this was leading to some delay in the screen of the app
coming up. In emulator, the coregrapher was putting logs about
frames being skipped. And earlier today when I tested the app on
a actual mobile, it took few seconds for the screen to show things,
after a initial white screen.

Now I have moved this initial load logic out of the FManFrags View
Create. Instead the MainActivities onCreate starts a co-routine
at the end, which calls the loadPath logic in a IO thread and once
its done, next it updates the SRcV on the Main thread from within
the co-routine.

This has avoided that dropped frames (180-500) log, when running
on the emulator (avd). Later have to check on a actual device.