FMan: Handle failed dirlist size (num of dir entries)
One may not have access to a subdir. In which case its listing will
fail and will be null, so we cant get size of the list. This is
taken care of now by setting it to 0.

TODO: Maybe change the fallback in this case to -1, so user knows
indirectly.