ShowSize: Setup sizeView, onBindVH control visibility
As user can change whether file/dirsize is shown or not at any time.
So sizeViews' font adjustment needs to be done always irrespective
of whether currently size is enabled or not.

Inturn onBindVH takes care of making the sizeView visible or not,
based on at that moment whether bGetSize is enabled or not.

As basic testing of the GUI logic is done. And also as user can
now control it, so by default I set bGetSize to false.