list views SelectedNoHELPEither: Will revert
As android was not giving any visible difference between a item
which is just in focus and one which is in focus as well as
activated, so tried using selected instead of activated. But even
this didnt help. SO WILL REVERT.

NOTE: The colors were different between Focus and Selected/Activated
but didnt help. Same with opacity.

NOTE: Have been using Activated and not selected, and again going
back to same, because activated is supposed to indicate longer term
selection. And also bcas neither selected nor activated is helping
wrt identifying selected/activated item which also has focus.