View: MultipleStates in State Selector list
Document doesnt seem to mention this clearly and or I didnt have
the patience to RTFM fully to see where if this info is.

Eitherway after giving up (ie not getting the desired result)
with other experiments to try and get the view to merge predefined
states where required automatically like for ex focus+activated.

Realised from the results that android doesnt auto merge the state
related drawables. So thought let me see if I can put more than one
state for a given drawable selector. And bingo (kulja simsim...)
it works.

THis is the initial experiment wrt same.