Main: settings/impGlobal to sharedPref, tempPref to bundle
Now the settings as well as the important global like curPath is
saved into sharedPreferences, so that even if the app is fully
stopped, It can restore these at any time in the future, even
across reboots.

However the transient state of currently selected/copied files list
is only saved in the outState bundle passed to saveInstanceState,
this ensures that if the app is restarted in the same session, like
when it regains foreground control of the device and end user, then
the end users actions are not lost. However if the app is fully
stopped and restarted as a new session/instance, then this now stale
info is not restored.