FMan: Bit more robust getVolumes, unmounted paths
getExternalFilesDirs may even return some paths which are null.
This can occur, if the storage volume/mount associated with that
path is currently unmounted or so.

Logic updated to take care of this. Also when at it now, it also
explicitly checks that the storage volume/mount associated with
its paths are mounted and available either in readwrite or readonly
mode.