Main: defaultPathString and fallback getDefaultVolume
As the curPath is now saved by the program when it quits, so when
it is started fresh, it will try to load this previously saved
path. But this fresh start may occur after the user has unmounted
say a sdcard, which is what the user might have been browsing,
when the application was quit.

So we can no longer assume that the defaultPathString loaded from
saved config during onCreate is valid. So inturn loadDefaultPath
has been updated to first try the saved path and inturn if it
fails, then fallback to getDefaultVolume, which currently is the
internal storage space alloted to the application and thus should
ideally be always present.

So also the showing of the currently loaded path in the GUI, has
been moved from onCreate to loadDefaultPath.